+++ src/Texture.cpp
@@ -19,7 +19,7 @@ Texture::Texture(const Utility::Path& path)
 	/////////////////////////////////////////////////////////////////////
 	//cpu load
 	//get raw file
-	void *data=NULL; uint len=0;
+	void *data=NULL; size_t len=0;
 	Application::instance()->loadData(path,data,len);
 	//load image
 	Image image;
@@ -30,7 +30,7 @@ Texture::Texture(const Utility::Path& path)
 	free(data);
 	//save width end height
 	width=image.width;
-	height=image.height;	
+	height=image.height;
 	//create a gpu texture
 	buildTexture(image.bytes,image.type);
 	//
@@ -47,12 +47,12 @@ Texture::Texture(const Vec4& floatColor,uint newWidth,uint newheight)
 	Image image(newWidth,newheight,4,true,Image::rgba(&(floatColor.x)));
 	//save width end height
 	width=newWidth;
-	height=newheight;	
+	height=newheight;
 	//create a gpu texture
 	buildTexture(image.bytes,image.type);
-}	
+}
 //send texture to gpu
-void Texture::buildTexture(void *data,uint type){	
+void Texture::buildTexture(void *data,uint type){
 	//gen gpu
 	//create an GPU texture
 	glGenTextures( 1, &gpuid );
@@ -69,7 +69,7 @@ void Texture::buildTexture(void *data,uint type){
 				 GL_UNSIGNED_BYTE,
 				 0);
 	//create mipmaps
-	glTexParameteri( GL_TEXTURE_2D, GL_GENERATE_MIPMAP, bMipmaps );	
+	glTexParameteri( GL_TEXTURE_2D, GL_GENERATE_MIPMAP, bMipmaps );
 	//send to GPU
     glTexSubImage2D( GL_TEXTURE_2D, 0, 0, 0,
 					 width,
@@ -145,7 +145,7 @@ Texture1D::Texture1D(const Utility::Path& path)
 	/////////////////////////////////////////////////////////////////////
 	//cpu load
 	//get raw file
-	void *data=NULL; uint len=0;
+	void *data=NULL; size_t len=0;
 	Application::instance()->loadData(path,data,len);
 	//load image
 	Image image;
@@ -155,7 +155,7 @@ Texture1D::Texture1D(const Utility::Path& path)
 	//free raw file
 	free(data);
 	//save width end height
-	width=image.width;	
+	width=image.width;
 	//create a gpu texture
 	buildTexture1D(image.bytes,image.type);
 	//
@@ -171,9 +171,9 @@ Texture1D::Texture1D(const Vec4& floatColor,uint newWidth)
 	width=newWidth;
 	//create a gpu texture
 	buildTexture1D(image.bytes,image.type);
-}	
+}
 //send texture to gpu
-void Texture1D::buildTexture1D(void *data,uint type){	
+void Texture1D::buildTexture1D(void *data,uint type){
 	//gen gpu
 	//create an GPU texture
 	glGenTextures( 1, &gpuid );
@@ -189,7 +189,7 @@ void Texture1D::buildTexture1D(void *data,uint type){
 				 GL_UNSIGNED_BYTE,
 				 0);
 	//send to GPU
-    glTexSubImage1D( GL_TEXTURE_1D, 0, 0, 
+    glTexSubImage1D( GL_TEXTURE_1D, 0, 0,
 					 width,
 					 type,
 					 GL_UNSIGNED_BYTE,
@@ -256,7 +256,7 @@ RenderTexture::RenderTexture(uint argwidth,uint argheight):Texture(),fboid(0),de
 	glBindTexture( GL_TEXTURE_2D, (GLuint)gpuid );
 	//save width end height
 	width=argwidth;
-	height=argheight;	
+	height=argheight;
 	//create a gpu texture
 	glTexImage2D(GL_TEXTURE_2D,
 				 0,
@@ -268,7 +268,7 @@ RenderTexture::RenderTexture(uint argwidth,uint argheight):Texture(),fboid(0),de
 				 GL_UNSIGNED_BYTE,
 				 0);
 	//create mipmaps
-	glTexParameteri( GL_TEXTURE_2D, GL_GENERATE_MIPMAP, bMipmaps );	
+	glTexParameteri( GL_TEXTURE_2D, GL_GENERATE_MIPMAP, bMipmaps );
 	//unbind texture
 	glBindTexture( GL_TEXTURE_2D, 0 );
 	//////////////////////////////////
@@ -278,10 +278,10 @@ RenderTexture::RenderTexture(uint argwidth,uint argheight):Texture(),fboid(0),de
 	//df buffer
 	glGenRenderbuffersEXT(1, &depthid);
     glBindRenderbufferEXT(GL_RENDERBUFFER_EXT, depthid);
-    glRenderbufferStorageEXT(GL_RENDERBUFFER_EXT, 
-							 GL_DEPTH_COMPONENT24, 
+    glRenderbufferStorageEXT(GL_RENDERBUFFER_EXT,
+							 GL_DEPTH_COMPONENT24,
 							 width,
-							 height);	
+							 height);
 	glBindRenderbufferEXT(GL_RENDERBUFFER_EXT, 0);
 	//////////////////////////////////
 	//get errors...
@@ -291,17 +291,17 @@ RenderTexture::RenderTexture(uint argwidth,uint argheight):Texture(),fboid(0),de
 	glGenFramebuffersEXT(1, &fboid);
 	glBindFramebufferEXT(GL_FRAMEBUFFER_EXT, fboid);
 	// Instruct openGL that we won't bind a color texture with the currently binded FBO
-	glFramebufferTexture2DEXT(GL_FRAMEBUFFER_EXT, 
+	glFramebufferTexture2DEXT(GL_FRAMEBUFFER_EXT,
 							  GL_COLOR_ATTACHMENT0_EXT ,
-							  GL_TEXTURE_2D, 
+							  GL_TEXTURE_2D,
 							  gpuid,
-							  0);	
+							  0);
     //-------------------------
     //Attach depth buffer to FBO
-    glFramebufferRenderbufferEXT(GL_FRAMEBUFFER_EXT, 
-								 GL_DEPTH_ATTACHMENT_EXT, 
-								 GL_RENDERBUFFER_EXT, 
-								 depthid);	
+    glFramebufferRenderbufferEXT(GL_FRAMEBUFFER_EXT,
+								 GL_DEPTH_ATTACHMENT_EXT,
+								 GL_RENDERBUFFER_EXT,
+								 depthid);
 	//set draw buffer
 	//glDrawBuffer(GL_COLOR_ATTACHMENT0_EXT);
 	//glReadBuffer(GL_COLOR_ATTACHMENT0_EXT);
@@ -310,7 +310,7 @@ RenderTexture::RenderTexture(uint argwidth,uint argheight):Texture(),fboid(0),de
 	if(status != GL_FRAMEBUFFER_COMPLETE_EXT){
 		DEBUG_MESSAGE("GL_FRAMEBUFFER_COMPLETE_EXT failed, CANNOT use FBO\n");
 	}
-	// disabilita fbo	
+	// disabilita fbo
 	glBindFramebufferEXT(GL_FRAMEBUFFER_EXT, 0);
 	//////////////////////////////////
 	//get errors...
@@ -318,14 +318,14 @@ RenderTexture::RenderTexture(uint argwidth,uint argheight):Texture(),fboid(0),de
 }
 //destructor
 RenderTexture::~RenderTexture(){
-	
+
 	//unload
 	DEBUG_ASSERT(gpuid);
     glDeleteRenderbuffersEXT(1, &depthid);
 	glDeleteFramebuffersEXT(1, &fboid);
 }
 //start draw
-void RenderTexture::enableRender(){	
+void RenderTexture::enableRender(){
 	//abilita fbo
 	glBindFramebufferEXT(GL_FRAMEBUFFER_EXT,fboid);
 }
@@ -344,17 +344,17 @@ void RenderTexture::draw(bool bindTexture){
 	glLoadIdentity();
 	//
 	static const float
-	size=1.0;	
+	size=1.0;
 	static const float
-	xyUV[]={ 
-			-size,-size,0.0,0.0, 
-			-size, size,0.0,1.0,  
-			 size,-size,1.0,0.0, 
+	xyUV[]={
+			-size,-size,0.0,0.0,
+			-size, size,0.0,1.0,
+			 size,-size,1.0,0.0,
 			 size, size,1.0,1.0
 	        };
 
 	glBindBuffer( GL_ARRAY_BUFFER, 0 );
-	glBindBuffer( GL_ELEMENT_ARRAY_BUFFER, 0 ); 
+	glBindBuffer( GL_ELEMENT_ARRAY_BUFFER, 0 );
 	//set texture
 	if(bindTexture) bind();
 	//set vertex
